+++ src/components/Section/ProposalVotingSectionTestCases.tsx
